+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "start": "cd src && nodemon app.js",
+    "start": "cd src && node app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
